requirements.txt
fastapi
uvicorn[standard]
PyMuPDF==1.22.1
opencv-python
numpy
python-multipart # For file uploads
ultralytics
Pillow==9.5.0
python-dotenv==1.0.0
pydantic==1.10.8
starlette==0.26.1
gradio==3.33.1
albumentations==1.3.0
tqdm # Used in detector for progress bars
pyyaml==6.0
onnx==1.14.0
onnxruntime==1.15.1
matplotlib==3.7.1
